Kernel/Storage: Rename ATA.h => Definitions.h
