+++ Kernel/Storage/ATA/AHCI/Port.cpp
@@ -13,8 +13,8 @@
 #include <Kernel/Memory/ScatterGatherList.h>
 #include <Kernel/Memory/TypedMapping.h>
 #include <Kernel/Storage/ATA/AHCI/Port.h>
-#include <Kernel/Storage/ATA/ATA.h>
 #include <Kernel/Storage/ATA/ATADiskDevice.h>
+#include <Kernel/Storage/ATA/Definitions.h>
 #include <Kernel/Storage/StorageManagement.h>
 #include <Kernel/WorkQueue.h>
